refactor: split portal into sections
